+++ where_to_go/lib/features/favorite/favorite_provider.dart
@@ -1,0 +1,15 @@
+import "package:riverpod_annotation/riverpod_annotation.dart";
+
+part "favorite_provider.g.dart";
+
+@riverpod
+class Favorite extends _$Favorite {
+  @override
+  bool build() {
+    return false; // stan początkowy
+  }
+
+  void toggle() {
+    state = !state;
+  }
+}
